Position and style hero text section of main
The hero text section is 80% of its container's width in order to get
its text to start wrapping sooner for neater presentation.

The text properties here are shared with the <p> tag present in the
form to maintain consistency. The reason for it being in the form and
not in the hero section is so that it hugs up against the form section
and not floating above it.